Refactor analytics data handling: relocate `getAnalyticsData` to `actions` and update usage in components.
